requirements.txt
# NPQsimple Python dependencies
# Tested on Python 3.11+. Minor versions are locked with ~= for stability while allowing patch updates.
# Omniboard is NOT a Python package; install via npm: `npm i -g omniboard`.

dash~=2.12
dash-bootstrap-components~=1.5
dash-daq~=0.6
plotly~=5.18

dnspython~=2.4
incense
ipdb~=0.13
joblib~=1.3
matplotlib~=3.8
numpy~=1.26
pandas~=2.2
pdf2image~=1.17
pillow~=10.2
pymongo~=4.6
python-dotenv~=1.0
sacred~=0.8.6
scipy~=1.12
seaborn~=0.13
sympy~=1.12
waitress~=2.1